* BolognaEAT:
  Author privilege is no longer required for viewing contents.
  Resolved more merge conflict.
  6a87e3882101d13d4495bb7acfa4a050f778c185
  5ea9b6e3aba9c4bad4d09390408be2d300a50208
  c50c65eca9ecbacbf881b0e2da7c5ecb7957dc9a
  19d6efd2ecf2c54f3c779761ecb4566c4ba9172f
  8305a7d3f94e5c180e594a94b7d8e4540c0d0c92
  4aa06d73fa3aef6c47148c80d141f0f2a26367bf
  af3a24e4a29de8acac57618acd091462cc13026c
  403e391167c88678d17c58b90f3b4eac5654b1e5
  removed config.inc.php file
  delete project file
  19fdf0ba901002aeb1ab8c03787b4d61d1d23721
  0a2ece34483d78c7e0eaf3201a8ffdd999792ff1
  9ebad121a0f55e8b19196def0529bc8c34d2c670
  555e691b7a1b632be1e526c1455eaa11b8e4c2b3
  53a81277c14f8cc362fb7c434570e5230f884f8d
  e947fd6ccd034d56c9658b29060d185c4431c408
  5d984e885009ba2246e207bc150676d1b5b1cf2f
  fe986cf704bd7418cd65f32640275e98a7e70e69
  f827dee744ebe415f1997da072f3555baf7e4c78
  2e544030fe3d2a79f9ef73c81bfc0a5a3690585a
  7e8d28927059486a50fab69228506ebf3db40c1f
  964df9905ef800ffe4deb4b632711583f501d417
  51a2085ac8a7054552fa041262085da0df292554
  bc0065164097779b48846a8d47077a06943d43e8
  93a840c239a1356d09ae1d832f337f121e69e003
  Add html code for label content
  Class label for successful save
  Code for correct  loading and code display
  Code for proper loading
  Code to write to db
  Code for displaying content added
  Enter code for table of possible layouts
  Enter code to handle the preview of the layout
  Entering class for the label page template
  Writing feature page template on DB
  Edit content central part
  New file for managing page template
  New file for managing page template
  Fixed bugs
  Inserting constant for the management of the template
  Changed the graphics interface of the layout template
  View Control tab structure template
  View Control tab and page layout templates
  Inserting constant for the management of the personal setting
  Inserting graphical elements for the management of the templates
  Inserting the values ​​of the control template
  Control value template
  Control value template
  Fixed bugs
  Moved layout template menu in ''edit content'' page
  Translated page template
  Translated the layout template
  Fixed bugs
  Added "create lesson" menu
  Fixed bug
  Added  the possibility to create a course with a specific structure or a wizard.
  Added  the possibility to create a course with a specific structure or a wizard.
  Fixed bugs
  Update dnd_themod/system/Models.js
  Update dnd_themod/system/Models.class.php
  Update dnd_themod/system/Themes.class.php
  Update dnd_themod/system/Structures.class.php
  Update dnd_themod/system/Themes.class.php
  Update dnd_themod/system/Themes.js
  Update dnd_themod/system/dnd_themod.css
  Update dnd_themod/system/models.css
  Update include/sidemenus/models.inc.php
  Update include/sidemenus/themes.inc.php
  Update include/side_menu.inc.php
  Fixed bug
  added file xml for templating
  Changed files txt (used for templating) in files xml
  Removed goals management
  Added new structure and new page templates
  Added "fill in the blank" test
  Create "fill in the blanks" test
  Minor fixes.
  Added the possibility to "Add top structure" in a existing lesson
  Removed .gitignore file.
  Minor fixes.
  Improved and updated structure template
  Integrated AContent with a system that support the use of three templates (layout template, page template and structure template) during authoring and editing of content.

+++ home/classes/ContentManager.class.php
@@ -39,6 +39,7 @@ class ContentManager
 	var $content_length;
 	var $dao;
 	var $contentDAO;
+	
 
 	/* constructor	*/
 	function ContentManager($course_id) {
@@ -48,6 +49,9 @@ class ContentManager
 		$this->course_id = $course_id;
 		$this->dao = new DAO();
 		$this->contentDAO = new ContentDAO();
+		
+		
+		
 		
 		/* x could be the ordering or even the content_id	*/
 		/* don't really need the ordering anyway.			*/
@@ -68,18 +72,24 @@ class ContentManager
 
 		$rows = $this->contentDAO->getContentByCourseID($this->course_id);
 		
+		
 		if (is_array($rows)) {
 			foreach ($rows as $row) {
+				
+				
+				
 				$num_sections++;
 				$_menu[$row['content_parent_id']][] = array('content_id'=> $row['content_id'],
 															'ordering'	=> $row['ordering'], 
 															'title'		=> htmlspecialchars($row['title']),
-															'content_type' => $row['content_type']);
+															'content_type' => $row['content_type'],
+															'optional' => $row['optional']);
 	
 				$_menu_info[$row['content_id']] = array('content_parent_id' => $row['content_parent_id'],
 														'title'				=> htmlspecialchars($row['title']),
 														'ordering'			=> $row['ordering'],
-														'content_type' => $row['content_type']);
+														'content_type' => $row['content_type'],
+														'optional' => $row['optional']);
 	
 				/* 
 				 * add test content asscioations
@@ -225,6 +235,7 @@ class ContentManager
 			$this->getContentPathRecursive($parent_id, $path);
 		}
 	}
+	
 
 	function addContent($course_id, $content_parent_id, $ordering, $title, $text, $keywords, 
 	                    $related, $formatting, $head = '', $use_customized_head = 0, 
@@ -240,12 +251,14 @@ class ContentManager
 		         WHERE ordering>=$ordering 
 		           AND content_parent_id=$content_parent_id 
 		           AND course_id=$_course_id";
+		 
 		$this->contentDAO->execute($sql);
-
+		
 		/* main topics all have minor_num = 0 */
 		$cid = $this->contentDAO->Create($_course_id, $content_parent_id, $ordering, 0, $formatting,
 		                          $keywords, '', $title, $text, $head, $use_customized_head,
 		                          $test_message, $content_type);
+		
 		return $cid;
 	}
 	
@@ -535,6 +548,7 @@ class ContentManager
 	 * @return an array of the next content node
 	 */
 	function getNextContent($content_id) {
+		//echo("dentro a get");
 		if (is_array($this->_menu_in_order))
 		{
 			// find out the location of the $content_id
@@ -562,8 +576,10 @@ class ContentManager
 				    	         'ordering'		=> $this->_menu_info[$this->_menu_in_order[$i]]['ordering'],
 				        	     'title'		=> $this->_menu_info[$this->_menu_in_order[$i]]['title']));
 			}
-		}
+		} 
+		
 		return NULL;
+		
 	}
 	
 	/* @See include/header.inc.php */
@@ -573,7 +589,7 @@ class ContentManager
 		$sequence_links = array();
 		
 		$first = $this->getNextContent(0); // get first
-		
+		//echo("TITOLO : ". $first['title']);
 		if ($_SESSION['prefs']['PREF_NUMBERING'] && $first) {
 			$first['title'] = $this->getNumbering($first['content_id']).' '.$first['title'];
 		}
@@ -707,6 +723,9 @@ function initContentMenu() {
 		{
 			echo "\n".'
   <div class="menuedit">
+  	<a href="'.$_base_path.'home/editor/edit_content_struct.php?_course_id='.$_course_id.'">
+      <img id="img_create_top_folder" src="'.$_base_path.'images/addstruct.gif" alt="Add top structure" title="Add top structure" style="border:0;height:1.2em" />
+    </a>'."\n".'
     <a href="'.$_base_path.'home/editor/edit_content_folder.php?_course_id='.$_course_id.'">
       <img id="img_create_top_folder" src="'.$_base_path.'images/mfolder.gif" alt="'._AT("add_top_folder").'" title="'._AT("add_top_folder").'" style="border:0;height:1.2em" />
     </a>'."\n".'
@@ -859,6 +878,7 @@ initContentMenu();
 			echo '<div id="folder'.$parent_id.$from.'">'."\n";
 			
 			foreach ($top_level as $garbage => $content) {
+				
 				$link = '';
 				//tests do not have content id
 				$content['content_id'] = isset($content['content_id']) ? $content['content_id'] : '';
@@ -917,8 +937,15 @@ initContentMenu();
 					
 					// instructors have privilege to delete content
 					if (isset($_current_user) && ($_current_user->isAuthor($this->course_id) || $_current_user->isAdmin()) && !isset($content['test_id']) && !Utility::isMobileTheme()) {
-						$link .= '<a href="'.$_base_path.'home/editor/delete_content.php?_cid='.$content['content_id'].'"><img src="'.TR_BASE_HREF.'images/x.gif" alt="'._AT("delete_content").'" title="'._AT("delete_content").'" style="border:0" height="10" /></a>';
+					//catia
+						if($content['optional'] == 1) 
+						//1 the content is optional
+							$link .= '<a href="'.$_base_path.'home/editor/delete_content.php?_cid='.$content['content_id'].'"><img src="'.TR_BASE_HREF.'images/x.gif" alt="'._AT("delete_content").'" title="'._AT("delete_content").'" style="border:0" height="10" /></a>';
+						else
+						//0 the content is mandatory
+							$link .= '<img style="margin-left:2px" src="'.$_base_path.'images/must.jpeg" title="mandatory page"/>';
 					}
+					
 				} 
 				else 
 				{ // current content page & nodes with content type "CONTENT_TYPE_FOLDER"
@@ -949,7 +976,14 @@ initContentMenu();
 						
 						// instructors have privilege to delete content
 						if (isset($_current_user) && ($_current_user->isAuthor($this->course_id) || $_current_user->isAdmin()) && !Utility::isMobileTheme()) {
-							$link .= '<a href="'.$_base_path.'home/editor/delete_content.php?_cid='.$content['content_id'].'"><img src="'.TR_BASE_HREF.'images/x.gif" alt="'._AT("delete_content").'" title="'._AT("delete_content").'" style="border:0" height="10" /></a>';
+						
+								//catia
+							if($content['optional'] == 1) 
+							//1 the content is optional
+								$link .= '<a href="'.$_base_path.'home/editor/delete_content.php?_cid='.$content['content_id'].'"><img src="'.TR_BASE_HREF.'images/x.gif" alt="'._AT("delete_content").'" title="'._AT("delete_content").'" style="border:0" height="10" /></a>';
+							else
+							//0 the content is mandatory
+								$link .= '<img src="'.$_base_path.'images/must.jpeg" title="Mandatory content" style="margin-left:2px;"/>';
 						}
 					}
 					else
